creación de cuádruplos correcta para el manejo de expresiones, aún falta resolver el problema el paréntesis y los condicionales. creación de cuádruplos correcta para, READ, WRITE E IF (sin else). corrección ocasional de gramática

+++ SemanticCube.py
@@ -6,7 +6,7 @@ semantic_cube["int"] = {
     "+" :  {"int" : "int", "float" : "float", "string" : "e"},
     "-" :  {"int" : "int", "float" : "float", "string" : "e"},
     "*" :  {"int" : "int", "float" : "float", "string" : "e"},
-    "/" :  {"int" : "int", "float" : "float", "string" : "e"},
+    "/" :  {"int" : "float", "float" : "float", "string" : "e"},
 
 
     #assignator
